feat: Add rudimentary plugin support
