+++ examples/projects/monorepo/packages/feature-account/codegen.yml
@@ -4,3 +4,5 @@ overwrite: true
 generates:
   '__generated__/baseTypes.ts':
     plugins: [typescript]
+  'introspection.json':
+    plugins: [introspection]
