+++ examples/projects/monorepo/packages/feature-account/introspection.json
@@ -1,0 +1,1 @@
+{"message":"introspection plugin is not implemented"}
