+++ internal/plugins/introspection.go
@@ -1,0 +1,5 @@
+package plugins
+
+func (p* PluginTask) Introspect() {
+    p.Output.WriteString("{\"message\":\"introspection plugin is not implemented\"}\n")
+}
